Add multi-page structure and branding
