+++ pages/__init__.py
@@ -1,0 +1,1 @@
+"""Page modules for PredictStream."""
