+++ static/logo.png
@@ -1,0 +1,1 @@
+placeholder logo
